Separate slicer passes and introduce StrategyOutput result type
Extract constant folding into onnx_fold.rs (793 LOC): fold_constant_nodes,
propagate_constants_with_shapes, propagate_constants, and 12 eval_*
functions. onnx_proto.rs reduced from 1854 to 1066 LOC.

Isolate tract integration into trace.rs (165 LOC): fold_and_trace_via_tract
extracted from onnx_slicer.rs, containing the tract dependency and
shape tracing callback. onnx_slicer.rs reduced from 1120 to 960 LOC.

Introduce StrategyOutput type wrapping ExecutionInfo with named tensor
outputs. Strategy modules (tiled, channel_split, dim_split) now return
outputs instead of mutating TensorStore directly. The dispatcher in
execute_slice stores returned tensors into the cache, establishing
a clear data flow boundary between strategy execution and state
management.
